split contexts into module for readability
